WIP on react node mongo starter

+++ src/components/Users/Users.js
@@ -12,18 +12,21 @@ const Users = () => {
   //DELETE AN USER
 
   const handleDeleteUser = (id) => {
-    const url = `http://localhost:5000/users/${id}`;
-    fetch(url, {
-      method: "DELETE",
-    })
-      .then((res) => res.json())
-      .then((data) => {
-        if (data.deletedCount > 0) {
-          alert("Deleted successfully");
-          const remainingUser = users.filter((user) => user._id !== id);
-          setUsers(remainingUser);
-        }
-      });
+    const proceed = window.confirm("Are you sure, you want to delete ?");
+    if (proceed) {
+      const url = `http://localhost:5000/users/${id}`;
+      fetch(url, {
+        method: "DELETE",
+      })
+        .then((res) => res.json())
+        .then((data) => {
+          if (data.deletedCount > 0) {
+            alert("Deleted successfully");
+            const remainingUser = users.filter((user) => user._id !== id);
+            setUsers(remainingUser);
+          }
+        });
+    }
   };
   return (
     <div>
